hackathonDB/s2t/auth/requirements.txt
google-cloud-texttospeech
google-cloud-speech
pyaudio
wave
google-cloud-dialogflow
google-cloud-dialogflow-cx

google-cloud-container
